requirements.txt
ib_insync>=0.9.86
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
rich>=13.0.0
jsonlines>=4.0.0
matplotlib>=3.7.0
pytest>=7.4.0
